[Power]: POWER_COMBO_POINTS

+++ src/server/scripts/Northrend/Ulduar/Ulduar/boss_yoggsaron.cpp
@@ -957,8 +957,8 @@ class boss_yoggsaron : public CreatureScript
                         if (Creature *pBrain = Creature::GetCreature((*me), instance->GetGuidData(DATA_YOGGSARON_BRAIN)))
                         {
                             pBrain->AI()->Reset();
-                            pBrain->AI()->DoAction((int32)ACTION_CHAMBER_ILLUSION);
-                            for (uint32 i = 0; i < RAID_MODE(4, 10); ++i)
+                            pBrain->AI()->DoAction(ACTION_CHAMBER_ILLUSION);
+                            for (int i = 0; i < RAID_MODE(4, 10); ++i)
                                 me->SummonCreature(NPC_PORTAL_CHAMBER, PortalPos[i], TEMPSUMMON_TIMED_DESPAWN, 10000);
                         }
                         break;
@@ -966,8 +966,8 @@ class boss_yoggsaron : public CreatureScript
                         if (Creature *pBrain = Creature::GetCreature((*me), instance->GetGuidData(DATA_YOGGSARON_BRAIN)))
                         {
                             pBrain->AI()->Reset();
-                            pBrain->AI()->DoAction((int32)ACTION_ICECROWN_ILLUSION);
-                            for (uint32 i = 0; i < RAID_MODE(4, 10); ++i)
+                            pBrain->AI()->DoAction(ACTION_ICECROWN_ILLUSION);
+                            for (int i = 0; i < RAID_MODE(4, 10); ++i)
                                 me->SummonCreature(NPC_PORTAL_ICECROWN, PortalPos[i], TEMPSUMMON_TIMED_DESPAWN, 10000);
 
                         }
@@ -976,8 +976,8 @@ class boss_yoggsaron : public CreatureScript
                         if (Creature *pBrain = Creature::GetCreature((*me), instance->GetGuidData(DATA_YOGGSARON_BRAIN)))
                         {
                             pBrain->AI()->Reset();
-                            pBrain->AI()->DoAction((int32)ACTION_STORMWIND_ILLUSION);
-                            for (uint32 i = 0; i < RAID_MODE(4, 10); ++i)
+                            pBrain->AI()->DoAction(ACTION_STORMWIND_ILLUSION);
+                            for (int i = 0; i < RAID_MODE(4, 10); ++i)
                                 me->SummonCreature(NPC_PORTAL_STORMWIND, PortalPos[i], TEMPSUMMON_TIMED_DESPAWN, 10000);
                         }
                         break;
